Decouple error message display from AccountStateViewModel
Extract toast/error infrastructure into a standalone ToastManager singleton
so composables and background services no longer need to reference
AccountStateViewModel just to show error dialogs.

- Add ToastManager.kt with ToastMsg sealed class hierarchy and MutableSharedFlow
- Remove toast methods and ToastMsg types from AccountStateViewModel
- Update DisplayErrorMessages() to take no parameter and collect from ToastManager
- Replace AmberListenerSingleton.accountStateViewModel field with direct ToastManager calls
- Update Amber.kt and BunkerRequestUtils.kt to use ToastManager directly
- Remove accountStateViewModel parameter from composables that only used it for toasts:
  NewApplicationScreen, ConfirmPinScreen, FeedbackScreen, IntentSingleEventHomeScreen,
  NewNsecBunkerScreen, EditConfigurationScreen, DefaultProfileRelaysScreen, EditRelaysDialog,
  IncomingRequestScreen, and update MainScreen callers accordingly
