+++ app/src/main/java/com/greenart7c3/nostrsigner/ui/IncomingRequestScreen.kt
@@ -36,7 +36,6 @@ fun IncomingRequestScreen(
     packageName: String?,
     applicationName: String?,
     account: Account,
-    accountStateViewModel: AccountStateViewModel,
     navController: NavController,
     onRemoveIntentData: (List<IntentData>, IntentResultType) -> Unit,
     onLoading: (Boolean) -> Unit,
@@ -70,7 +69,6 @@ fun IncomingRequestScreen(
                 applicationName = applicationName,
                 intentData = intents.first(),
                 account = account,
-                accountStateViewModel = accountStateViewModel,
                 onRemoveIntentData = onRemoveIntentData,
                 onLoading = onLoading,
             )
